Support testing multiple subseries per series in rankAndReport
Adds an optional parameter to rankAndReport which allows repating
the evaluation of the scoring function by taking subseries from the
end n times. The best one per full series is then reported.

Adds support for marking a candle in the chart in the webapp. This
enables rankAndReport to point to the subseries where the high score
was scored.