Remove RawCandle type
by renaming it to Candle and removing the old Candle type.
It just caused confusion, because RawCandles were the ones you
want to store to DB to avoid extra data, but then you easily forgot
to convert RawCandle to Candle where needed. The additional props
can be retrieved with functions when needed.