Remove provider wrapper from strategy arg of backtest()
The reason for this provider function was to make it less
likely to reuse the same strategy instance on multiple
backtests, which would cause issues if the strategy is
stateful (e.g. holding indicator values in a closure).

But it doesn't really solve anything, because users can
as well use a provider like
() => strat
instead of
() => createNewStrat()